+++ Source/Tools/JsonDocument.cpp
@@ -1,0 +1,3 @@
+#include "../AliveLibCommon/stdafx_common.h"
+#include "JsonDocument.hpp"
+
