+++ Source/Tools/alive_api.cpp
@@ -9,470 +9,19 @@
 
 #include <iostream>
 #include "magic_enum.hpp"
-
-#include <jsonxx/jsonxx.h>
+#include "JsonDocument.hpp"
+#include "AETlvs.hpp"
+#include "TypesCollection.hpp"
 #include <gmock/gmock.h>
 #include <type_traits>
 #include <typeindex>
 
-#define NAME_OF( v ) #v
 
 bool RunningAsInjectedDll()
 {
     return false;
 }
 
-struct PropertyTypeInfo
-{
-    std::string mMappedName;
-    std::string mTypeName;
-};
-using TFieldToNameAndTypeMapping = std::map<std::string, PropertyTypeInfo>;
-
-
-struct CollisionObject
-{
-    int mX1 = 0;
-    int mY1 = 0;
-    int mX2 = 0;
-    int mY2 = 0;
-    // TODO: Other properties
-
-    jsonxx::Object ToJsonObject() const
-    {
-        jsonxx::Object obj;
-
-        obj << "x1" << mX1;
-        obj << "y1" << mY1;
-
-        obj << "x2" << mX2;
-        obj << "y2" << mY2;
-
-        return obj;
-    }
-};
-
-struct ObjectStructureBasicType
-{
-    std::string mName;
-    int mMinValue = 0;
-    int mMaxValue = 0;
-
-    void AddToJsonArray(jsonxx::Array& array) const
-    {
-        jsonxx::Object maxRangeObject;
-        maxRangeObject << "min_value" << mMinValue;
-        maxRangeObject << "max_value" << mMaxValue;
-        array << mName << maxRangeObject;
-    }
-};
-
-struct ObjectStructurePropertyEnum
-{
-    std::string mName;
-    std::vector<std::string> mValues;
-
-    void AddJsonArray(jsonxx::Object& obj) const
-    {
-        jsonxx::Array valuesArray;
-        for (const auto& value : mValues)
-        {
-            valuesArray << value;
-        }
-        obj << mName << valuesArray;
-    }
-};
-
-struct MapObjectProperty
-{
-public:
-    bool mIsString = false;
-    int mIntValue = 0;
-    std::string mName;
-    std::string mStringValue;
-
-    void AddToJsonObject(jsonxx::Object& obj) const
-    {
-        if (mIsString)
-        {
-            obj << mName << mStringValue;
-        }
-        else
-        {
-            obj << mName << mIntValue;
-        }
-    }
-};
-
-struct Property
-{
-    std::string mName;
-    std::string mType;
-};
-
-struct ObjectStructure
-{
-    ObjectStructure() = default;
-    explicit ObjectStructure(const std::string& name) : mName(name) { }
-
-    std::string mName;
-    std::vector<Property> mProperties;
-
-    void AddProps(const TFieldToNameAndTypeMapping& propsToAdd)
-    {
-        for (const auto& [key, value] : propsToAdd)
-        {
-            Property property;
-            property.mName = value.mMappedName;
-            property.mType = value.mTypeName;
-            mProperties.push_back(property);
-        }
-    }
-
-    void AddToJsonArray(jsonxx::Array& array) const
-    {
-        jsonxx::Object obj;
-        obj << "name" << mName;
-        jsonxx::Array enumAndBasicTypePropertiesArray;
-        for (const auto& property : mProperties)
-        {
-            obj << property.mName << property.mType;
-        }
-        obj << "enum_and_basic_type_properties" << enumAndBasicTypePropertiesArray;
-        array << obj;
-    }
-
-};
-
-struct MapObject
-{
-    int mXPos = 0;
-    int mYPos = 0;
-    int mWidth = 0;
-    int mHeight = 0;
-    std::string mName;
-    std::string mObjectStructuresType;
-    std::vector<MapObjectProperty> mProperties;
-
-    jsonxx::Object ToJsonObject() const
-    {
-        jsonxx::Object obj;
-
-        obj << "xpos" << mXPos;
-        obj << "ypos" << mYPos;
-        obj << "width" << mWidth;
-        obj << "height" << mHeight;
-        obj << "name" << mName;
-        obj << "object_structures_type" << mObjectStructuresType;
-
-        jsonxx::Object propertiesObject;
-        for (const auto& property : mProperties)
-        {
-            property.AddToJsonObject(propertiesObject);
-        }
-        obj << "properties" << propertiesObject;
-
-        return obj;
-    }
-};
-
-struct CameraObject
-{
-    std::string mName;
-    int mId = 0;
-    int mX = 0;
-    int mY = 0;
-    std::vector<MapObject> mMapObjects;
-
-    jsonxx::Object ToJsonObject() const
-    {
-        jsonxx::Object obj;
-
-        obj << "name" << mName;
-        obj << "x" << mX;
-        obj << "y" << mY;
-        obj << "id" << mId;
-
-        jsonxx::Array mapObjectsArray;
-        for (const auto& mapObject : mMapObjects)
-        {
-            mapObjectsArray << mapObject.ToJsonObject();
-        }
-
-        obj << "map_objects" << mapObjectsArray;
-
-        return obj;
-    }
-};
-
-
-
-const std::map<std::string, std::string> kTypeNames =
-{
-    { NAME_OF(Path_Hoist), "Hoist" },
-    { NAME_OF(Path_Hoist::Type), "Enum_HoistType" },
-    { NAME_OF(Path_Hoist::EdgeType), "Enum_HoistEdgeType" },
-    { NAME_OF(Path_Hoist::Scale), "Enum_HoistScale" },
-
-    { NAME_OF(Path_Edge), "Edge" },
-    { NAME_OF(Path_Edge::Type), "Enum_EdgeType" },
-    { NAME_OF(Path_Edge::Scale), "Enum_EdgeScale" },
-
-};
-
-const std::map<std::string, std::string> kEnum_HoistScale =
-{
-    { NAME_OF(Path_Hoist::Scale::eFull), "full" },
-    { NAME_OF(Path_Hoist::Scale::eHalf), "half" },
-};
-
-const std::map<std::string, std::string> kEnum_HoistType =
-{
-    { NAME_OF(Path_Hoist::Type::eNextFloor), "next_floor" },
-    { NAME_OF(Path_Hoist::Type::eNextEdge), "next_edge" },
-    { NAME_OF(Path_Hoist::Type::eOffScreen), "off_screen" },
-};
-
-const std::map<std::string, std::string> kEnum_HoistEdgeType =
-{
-    { NAME_OF(Path_Hoist::EdgeType::eLeft), "left" },
-    { NAME_OF(Path_Hoist::EdgeType::eRight), "right" },
-    { NAME_OF(Path_Hoist::EdgeType::eBoth), "both" },
-};
-
-const std::map<std::string, std::string> kEnum_EdgeType =
-{
-    { NAME_OF(Path_Edge::EdgeType::eLeft), "left" },
-    { NAME_OF(Path_Edge::EdgeType::eRight), "right" },
-    { NAME_OF(Path_Edge::EdgeType::eBoth), "both" },
-};
-
-const std::map<std::string, std::string> kEnum_EdgeScale =
-{
-    { NAME_OF(Path_Edge::Scale::eFull), "full" },
-    { NAME_OF(Path_Edge::Scale::eHalf), "half" },
-};
-
-struct StructuresAndTypes
-{
-    void CreateFromJson(const std::string& /*json*/)
-    {
-
-    }
-
-    void CreateCurrentVersion()
-    {
-        AddBasicTypes();
-        Add_Path_Hoist();
-        Add_Path_Edge();
-    }
-
-    template<class T>
-    static std::vector<std::string> AsStringVec(const T& data)
-    {
-        std::vector<std::string> ret;
-        for (const auto& item : data)
-        {
-            ret.push_back(std::string(magic_enum::enum_name(item)));
-        }
-        return ret;
-    }
-
-    void AddBasicTypes()
-    {
-        AddBasicType("Bool", 0, 1);
-        AddBasicType("Byte", std::numeric_limits<BYTE>::min(), std::numeric_limits<BYTE>::max());
-        AddBasicType("Int32", std::numeric_limits<int>::min(), std::numeric_limits<int>::max());
-        AddBasicType("Int16", std::numeric_limits<short>::min(), std::numeric_limits<short>::max());
-    }
-
-    void AddBasicType(const std::string& name, int minVal, int maxVal)
-    {
-        ObjectStructureBasicType basicType;
-        basicType.mName = name;
-        basicType.mMinValue = minVal;
-        basicType.mMaxValue = maxVal;
-        mBasicTypes.push_back(basicType);
-    }
-
-    template<class T>
-    static std::string Value(const T& map, const std::string& key)
-    {
-        auto it = map.find(key);
-        if (it == std::end(map))
-        {
-            abort();
-        }
-        return it->second;
-    }
-
-    static std::string TypeName(const std::string& typeName)
-    {
-        if (typeName == NAME_OF(BYTE))
-        {
-            return "Byte";
-        }
-        else if (typeName == NAME_OF(bool))
-        {
-            return "Bool";
-        }
-        abort();
-    }
-
-    void Add_Path_Hoist()
-    {
-        const TFieldToNameAndTypeMapping kPath_Hoist =
-        {
-            { NAME_OF(Path_Hoist::field_10_type), {"HoistType", Value(kTypeNames, NAME_OF(Path_Hoist::Type)) } },
-            { NAME_OF(Path_Hoist::field_12_edge_type), {"EdgeType", Value(kTypeNames, NAME_OF(Path_Hoist::EdgeType)) } },
-            { NAME_OF(Path_Hoist::field_14_id), {"Id", TypeName(NAME_OF(BYTE)) } },
-            { NAME_OF(Path_Hoist::field_16_scale), {"Scale", Value(kTypeNames, NAME_OF(Path_Hoist::Scale)) } },
-        };
-
-        auto& type = AddStructure(Value(kTypeNames, NAME_OF(Path_Hoist)));
-        type.AddProps(kPath_Hoist);
-        AddEnum(Value(kTypeNames, NAME_OF(Path_Hoist::Type)), kEnum_HoistType);
-        AddEnum(Value(kTypeNames, NAME_OF(Path_Hoist::EdgeType)), kEnum_HoistEdgeType);
-        AddEnum(Value(kTypeNames, NAME_OF(Path_Hoist::Scale)), kEnum_HoistScale);
-    }
-
-    void Add_Path_Edge()
-    {
-        const TFieldToNameAndTypeMapping kPath_Edge =
-        {
-            { NAME_OF(Path_Edge::field_10_type), {"EdgeType", Value(kTypeNames, NAME_OF(Path_Edge::Type)) } },
-            { NAME_OF(Path_Edge::field_12_can_grab), {"CanGrab",  TypeName(NAME_OF(bool)) } },
-            { NAME_OF(Path_Edge::field_14_scale), {"Scale", Value(kTypeNames, NAME_OF(Path_Edge::Scale)) } },
-        };
-        auto& type = AddStructure(Value(kTypeNames, NAME_OF(Path_Edge)));
-        type.AddProps(kPath_Edge);
-        AddEnum(Value(kTypeNames, NAME_OF(Path_Edge::Type)), kEnum_EdgeType);
-        AddEnum(Value(kTypeNames, NAME_OF(Path_Edge::Scale)), kEnum_EdgeScale);
-    }
-
-
-    ObjectStructure& AddStructure(const std::string& typeName)
-    {
-        ObjectStructure structure;
-        structure.mName = typeName;
-        mObjectStructures.push_back(structure);
-
-        for (auto& item : mObjectStructures)
-        {
-            if (item.mName == typeName)
-            {
-                return item;
-            }
-        }
-        abort();
-    }
-
-    void AddEnum(const std::string& enumName, const std::map<std::string, std::string>& values)
-    {
-        ObjectStructurePropertyEnum newEnum;
-        newEnum.mName = enumName;
-        for (const auto& [key, value] : values)
-        {
-            newEnum.mValues.push_back(value);
-        }
-        mEnums.push_back(newEnum);
-    }
-
-    std::vector<ObjectStructure> mObjectStructures;
-    std::vector<ObjectStructureBasicType> mBasicTypes;
-    std::vector<ObjectStructurePropertyEnum> mEnums;
-};
-
-class JsonDocument
-{
-public:
-    int mVersion = 0;
-    std::string mGame;
-
-    std::string mPathBnd;
-    int mPathId = 0;
-
-    int mXGridSize = 0;
-    int mXSize = 0;
-
-    int mYGridSize = 0;
-    int mYSize = 0;
-
-    std::vector<CollisionObject> mCollisions;
-    std::vector<CameraObject> mCameras;
-
-    StructuresAndTypes mStructureAndTypes;
-
-    void Load(const std::string& fileName)
-    {
-
-    }
-
-    void Save(const std::string& fileName) const
-    {
-        jsonxx::Object rootObject;
-
-        rootObject << "api_version" << mVersion;
-
-        rootObject << "game" << mGame;
-
-        jsonxx::Object rootMapObject;
-        rootMapObject << "path_bnd" <<  mPathBnd;
-        rootMapObject << "path_id" << mPathId;
-
-        rootMapObject << "x_grid_size" << mXGridSize;
-        rootMapObject << "x_size" << mXSize;
-
-        rootMapObject << "y_grid_size" << mYGridSize;
-        rootMapObject << "y_size" << mYSize;
-
-        jsonxx::Array collisionsArray;
-        for (const auto& item : mCollisions)
-        {
-            collisionsArray << item.ToJsonObject();
-        }
-        rootMapObject << "collisions" << collisionsArray;
-
-        jsonxx::Array cameraArray;
-        for (const auto& camera : mCameras)
-        {
-            cameraArray << camera.ToJsonObject();
-        }
-        rootMapObject << "cameras" << cameraArray;
-
-        jsonxx::Object objectStructurePropertiesObject;
-        for (const auto& propertyEnum : mStructureAndTypes.mEnums)
-        {
-            propertyEnum.AddJsonArray(objectStructurePropertiesObject);
-        }
-        rootMapObject << "object_structure_property_enums" << objectStructurePropertiesObject; 
-
-        jsonxx::Array objectStructureBasicTypes;
-        for (const auto& basicType : mStructureAndTypes.mBasicTypes)
-        {
-            basicType.AddToJsonArray(objectStructureBasicTypes);
-        }
-        rootMapObject << "object_structure_property_basic_types" << objectStructureBasicTypes;
-
-        jsonxx::Array objectStructures;
-        for (const auto& objectStructure : mStructureAndTypes.mObjectStructures)
-        {
-            objectStructure.AddToJsonArray(objectStructures);
-        }
-        rootMapObject << "object_structures" << objectStructures;
-
-        rootObject << "map" << rootMapObject;
-
-        std::ofstream s(fileName.c_str());
-        if (s)
-        {
-            s << rootObject.json();
-        }
-
-    }
-
-
-};
 
 namespace AliveAPI
 {
@@ -480,20 +29,10 @@ namespace AliveAPI
     // upgrade step that converts from the last version to the current.
     constexpr int kApiVersion = 1;
 
-    int GetApiVersion()
+    [[nodiscard]] int GetApiVersion()
     {
         return kApiVersion;
     }
-
-    struct PathInfo
-    {
-        int mGridWidth = 0;
-        int mWidth = 0;
-        int mGridHeight = 0;
-        int mHeight = 0;
-        unsigned int mObjectOffset = 0;
-        unsigned int mIndexTableOffset = 0;
-    };
 
     struct PathBND
     {
@@ -512,7 +51,7 @@ namespace AliveAPI
         return fileContent;
     }
 
-    static PathBND OpenPathBnd(const std::string& inputLvlFile, int* pathId)
+    [[nodiscard]] static PathBND OpenPathBnd(const std::string& inputLvlFile, int* pathId)
     {
         PathBND ret = {};
 
@@ -600,28 +139,82 @@ namespace AliveAPI
         return ret;
     }
 
-    Result ExportPathBinaryToJson(const std::string& /*jsonOutputFile*/, const std::string& outputLvlFile, int pathResourceId)
+    [[nodiscard]] Result ExportPathBinaryToJson(const std::string& /*jsonOutputFile*/, const std::string& inputLvlFile, int pathResourceId)
     {
+        TypesCollection globalTypes;
+
         ResourceManager::Init_49BCE0();
-        Result ret = {};
-        PathBND pathBnd = OpenPathBnd(outputLvlFile, &pathResourceId);
+        AliveAPI::Result ret = {};
+        AliveAPI::PathBND pathBnd = AliveAPI::OpenPathBnd(inputLvlFile, &pathResourceId);
         ret.mResult = pathBnd.mResult;
 
+        BYTE* pPathData = pathBnd.mFileData.data();
+        const int* indexTable = reinterpret_cast<const int*>(pPathData + pathBnd.mPathInfo.mIndexTableOffset);
+        int idx = 0;
+
+        std::set<TlvTypes> usedTypes;
+
+        for (int x = 0; x < pathBnd.mPathInfo.mWidth; x++)
+        {
+            for (int y = 0; y < pathBnd.mPathInfo.mHeight; y++)
+            {
+                const int objectTableIdx = indexTable[idx++];
+                if (objectTableIdx == -1 || objectTableIdx >= 0x100000)
+                {
+                    continue;
+                }
+
+                BYTE* ptr = pPathData + objectTableIdx + pathBnd.mPathInfo.mObjectOffset;
+                AO::Path_TLV* pPathTLV = reinterpret_cast<AO::Path_TLV*>(ptr);
+                pPathTLV->RangeCheck();
+                if (pPathTLV->field_4_type <= 0x100000 && pPathTLV->field_2_length <= 0x2000u && pPathTLV->field_8 <= 0x1000000)
+                {
+                    for (;;)
+                    {
+                        // End of TLV list for current camera
+                        if (pPathTLV->field_0_flags.Get(AO::TLV_Flags::eBit3_End_TLV_List))
+                        {
+                            break;
+                        }
+
+                        usedTypes.insert(static_cast<TlvTypes>(pPathTLV->field_4_type));
+                        switch (pPathTLV->field_4_type)
+                        {
+                        case AO::TlvTypes::Hoist_3:
+                        {
+                            AETlvs::Path_Hoist hoist(globalTypes, (Path_TLV*)pPathTLV);
+                            hoist.InstanceToJson(globalTypes);
+
+                        }
+                        break;
+                        }
+
+                        pPathTLV = AO::Path_TLV::Next_446460(pPathTLV);
+                        pPathTLV->RangeCheck();
+                    }
+                }
+            }
+        }
+
+        JsonDocument doc;
+        doc.mGame = "AE";
+        //doc.SaveAE(1, pathBnd.mPathInfo, pathBnd.mFileData, "lols.json");
+
 
         return {};
     }
 
-    Result UpgradePathJson(const std::string& /*jsonFile*/)
+    [[nodiscard]] Result UpgradePathJson(const std::string& /*jsonFile*/)
     {
         return {};
     }
 
-    Result ImportPathJsonToBinary(const std::string& /*jsonInputFile*/, const std::string& /*inputLvlFile*/, int /*pathResourceId*/, const std::vector<std::string>& /*lvlResourceSources*/)
+    [[nodiscard]] Result ImportPathJsonToBinary(const std::string& /*jsonInputFile*/, const std::string& /*outputLvlFile*/, int /*pathResourceId*/, const std::vector<std::string>& /*lvlResourceSources*/)
     {
         return {};
     }
 
-    EnumeratePathsResult EnumeratePaths(const std::string& inputLvlFile)
+    [[nodiscard]] EnumeratePathsResult EnumeratePaths(const std::string& inputLvlFile)
     {
         ResourceManager::Init_49BCE0();
         EnumeratePathsResult ret = {};
@@ -642,550 +235,13 @@ const std::map<std::string, AO::TlvTypes> kObjectNameToAoTlv =
     { "Hoist", AO::TlvTypes::Hoist_3 },
 };
 
-class ITypeBase
-{
-public:
-    explicit ITypeBase(const std::string& typeName) : mName(typeName)
-    {
-
-    }
-
-    std::string Name() const
-    {
-        return mName;
-    }
-
-    virtual std::type_index TypeIndex() const = 0;
-    virtual bool IsBasicType() const = 0;
-    virtual void ToJson(jsonxx::Object& obj) const = 0;
-private:
-    std::string mName;
-};
-
-template<class T>
-class BasicType : public ITypeBase
-{
-public:
-    explicit BasicType(const std::string& typeName, int minVal, int maxVal)
-        : ITypeBase(typeName), mTypeIndex(typeid(T)), mMinVal(minVal), mMaxVal(maxVal)
-    {
-
-    }
-
-    std::type_index TypeIndex() const override
-    {
-        return mTypeIndex;
-    }
-
-    bool IsBasicType() const override
-    {
-        return true;
-    }
-
-    void ToJson(jsonxx::Object& obj) const override
-    {
-        jsonxx::Object ret;
-
-        ret << "min_value" << mMinVal;
-        ret << "max_value" << mMaxVal;
-
-        obj << Name() << ret;
-    }
-
-private:
-    std::type_index mTypeIndex;
-    int mMinVal = 0;
-    int mMaxVal = 0;
-};
-
-template<class T>
-class EnumType : public ITypeBase
-{
-public:
-    explicit EnumType(const std::string& typeName)
-        : ITypeBase(typeName), mTypeIndex(typeid(T))
-    {
-
-    }
-
-    void Add(T enumValue, const std::string& name)
-    {
-        mMapping[enumValue] = name;
-    }
-
-    std::type_index TypeIndex() const override
-    {
-        return mTypeIndex;
-    }
-
-    T ValueFromString(const std::string& valueString) const
-    {
-        for (const auto [key, value] : mMapping)
-        {
-            if (value == valueString)
-            {
-                return key;
-            }
-        }
-        abort();
-    }
-
-    std::string ValueToString(T valueToFind) const
-    {
-        for (const auto [key, value] : mMapping)
-        {
-            if (key == valueToFind)
-            {
-                return value;
-            }
-        }
-        abort();
-    }
-
-    bool IsBasicType() const override
-    {
-        return false;
-    }
-
-    void ToJson(jsonxx::Object& obj) const override
-    {
-        jsonxx::Array enumVals;
-        for (const auto& [key, value] : mMapping)
-        {
-            enumVals << value;
-        }
-        obj << Name() << enumVals;
-    }
-
-private:
-    std::map<T, std::string> mMapping;
-    std::type_index mTypeIndex;
-};
-
-class TypesCollection
-{
-public:
-    TypesCollection()
-    {
-        AddBasicType<BYTE>("Byte", 0, 255);
-        AddBasicType<short>("UInt16", 0, 65535);
-
-        AddEnum<::Path_Hoist::Type>("Enum_HoistType",
-            {
-                {::Path_Hoist::Type::eNextEdge, "next_edge"},
-                {::Path_Hoist::Type::eNextFloor, "next_floor"},
-                {::Path_Hoist::Type::eOffScreen, "off_screen"},
-            });
-
-        AddEnum<::Path_Hoist::EdgeType>("Enum_HoistEdgeType",
-            {
-                {::Path_Hoist::EdgeType::eBoth, "both"},
-                {::Path_Hoist::EdgeType::eLeft, "left"},
-                {::Path_Hoist::EdgeType::eRight, "right"},
-            });
-
-        AddEnum<::Path_Hoist::Scale>("Enum_HoistScale",
-            {
-                {::Path_Hoist::Scale::eFull, "full"},
-                {::Path_Hoist::Scale::eHalf, "half"}
-            });
-    }
-
-    jsonxx::Object EnumsToJson() const
-    {
-        jsonxx::Object ret;
-        for (const auto& basicType : mTypes)
-        {
-            if (!basicType->IsBasicType())
-            {
-                basicType->ToJson(ret);
-            }
-        }
-        jsonxx::Object t;
-        t << "object_structure_property_enums" << ret;
-        return t;
-    }
-
-    jsonxx::Object BasicTypesToJson() const
-    {
-        jsonxx::Object ret;
-        for (const auto& basicType : mTypes)
-        {
-            if (basicType->IsBasicType())
-            {
-                basicType->ToJson(ret);
-            }
-        }
-        jsonxx::Object t;
-        t << "object_structure_property_basic_types" << ret;
-        return t;
-    }
-
-    std::string TypeName(std::type_index typeIndex) const
-    {
-        for (const auto& e : mTypes)
-        {
-            if (e->TypeIndex() == typeIndex)
-            {
-                return e->Name();
-            }
-        }
-        return "";
-    }
-
-    template<class T>
-    std::string TypeName() const
-    {
-        for (const auto& e : mTypes)
-        {
-            if (e->TypeIndex() == typeid(T))
-            {
-                return e->Name();
-            }
-        }
-        return "";
-    }
-
-    template<class T>
-    struct EnumPair
-    {
-        T mEnumValue;
-        std::string mName;
-    };
-
-    template<class T>
-    auto AddEnum(const std::string& enumName, const std::vector<EnumPair<T>>& enumItems)
-    {
-        EnumType<T>* ret = nullptr;
-        if (!TypeName<T>().empty())
-        {
-            // Type already exists
-            return ret;
-        }
-
-        auto newEnum = std::make_unique<EnumType<T>>(enumName);
-        for (const auto& enumItem : enumItems)
-        {
-            newEnum->Add(enumItem.mEnumValue, enumItem.mName);
-        }
-        ret = newEnum.get();
-        mTypes.push_back(std::move(newEnum));
-        return ret;
-    }
-
-    template<class T>
-    T EnumValueFromString(const std::string& enumTypeName, const std::string& enumValueString)
-    {
-        for (const auto& e : mTypes)
-        {
-            if (e->Name() == enumTypeName)
-            {
-               return static_cast<EnumType<T>*>(e.get())->ValueFromString(enumValueString);
-            }
-        }
-        abort();
-    }
-
-    template<class T>
-    std::string EnumValueToString(T enumValue)
-    {
-        for (const auto& e : mTypes)
-        {
-            if (e->TypeIndex() == typeid(T))
-            {
-                return static_cast<EnumType<T>*>(e.get())->ValueToString(enumValue);
-            }
-        }
-        abort();
-    }
-
-    template<class T>
-    auto AddBasicType(const std::string& typeName, int minVal, int maxVal)
-    {
-        BasicType<T>* ret = nullptr;
-        if (!TypeName<T>().empty())
-        {
-            // Type already exists
-            return ret;
-        }
-
-        auto newType = std::make_unique<BasicType<T>>(typeName, minVal, maxVal);
-        ret = newType.get();
-        mTypes.push_back(std::move(newType));
-        return ret;
-    }
-
-
-private:
-    std::vector<std::unique_ptr<ITypeBase>> mTypes;
-};
-
-class IMapObject
-{
-public:
-    IMapObject(const std::string& typeName, Path_TLV& tlv)
-        : mName(typeName), mTlv(tlv)
-    {
-
-    }
-
-    virtual ~IMapObject() {}
-
-    std::string Name() const
-    {
-        return mName;
-    }
-
-    void AddProperty(const std::string& name, const std::string& typeName, void* key)
-    {
-        mInfo[key] = { name, typeName };
-    }
-
-    jsonxx::Object PropertiesToJson() const
-    {
-        jsonxx::Object ret;
-        for (const auto& [key, value] : mInfo)
-        {
-            ret << value.mName << value.mTypeName;
-        }
-        return ret;
-    }
-
-    std::string PropName(void* key) const
-    {
-        auto it = mInfo.find(key);
-        if (it == std::end(mInfo))
-        {
-            abort();
-        }
-        return it->second.mName;
-    }
-
-    std::string PropType(void* key) const
-    {
-        auto it = mInfo.find(key);
-        if (it == std::end(mInfo))
-        {
-            abort();
-        }
-        return it->second.mTypeName;
-    }
-
-    template<class T>
-    void ReadEnumValue(TypesCollection& types, T& field, jsonxx::Object& properties)
-    {
-        const std::string propName = PropName(&field);
-        const std::string propType = PropType(&field);
-        const std::string jsonValue = properties.get<std::string>(propName);
-        field = types.EnumValueFromString<T>(propType, jsonValue);
-    }
-
-    template<class T>
-    void WriteEnumValue(TypesCollection& types, jsonxx::Object& properties, T& field)
-    {
-        properties << PropName(&field) << types.EnumValueToString<T>(field);
-    }
-
-    template<class T>
-    void ReadBasicType(T& field, jsonxx::Object& properties)
-    {
-        field = properties.get<jsonxx::Number>(PropName(&field));
-    }
-
-    template<class T>
-    void WriteBasicType(T& field, jsonxx::Object& properties)
-    {
-        properties << PropName(&field) << field;
-    }
-
-    template<class T>
-    void InstanceFromJsonBase(T& data, jsonxx::Object& obj)
-    {
-        mDescription = obj.get<std::string>("name");
-
-        data.field_8_top_left.field_0_x = obj.get<jsonxx::Number>("xpos");
-        data.field_8_top_left.field_2_y = obj.get<jsonxx::Number>("ypos");
-        data.field_C_bottom_right.field_0_x = obj.get<jsonxx::Number>("width");
-        data.field_C_bottom_right.field_2_y = obj.get<jsonxx::Number>("height");
-    }
-
-    template<class T>
-    void InstanceToJsonBase(T& data,jsonxx::Object& ret)
-    {
-        ret << "name" << mDescription;
-
-        ret << "xpos" << static_cast<int>(data.field_8_top_left.field_0_x);
-        ret << "ypos" << static_cast<int>(data.field_8_top_left.field_2_y);
-        ret << "width" << static_cast<int>(data.field_C_bottom_right.field_0_x);
-        ret << "height" << static_cast<int>(data.field_C_bottom_right.field_2_y);
-
-        ret << "object_structures_type" << Name();
-    }
-
-
-    jsonxx::Object StructureToJson()
-    {
-        jsonxx::Object ret;
-        ret << "name" << Name();
-        ret << "enum_and_basic_type_properties" << PropertiesToJson();
-        return ret;
-    }
-
-    void InstanceFromJson(TypesCollection& types, jsonxx::Object& obj)
-    {
-        InstanceFromJsonBase(mTlv, obj);
-        jsonxx::Object properties = obj.get<jsonxx::Object>("properties");
-        PropertiesFromJson(types, properties);
-    }
-
-    virtual void PropertiesFromJson(TypesCollection& types, jsonxx::Object& properties) = 0;
-
-    jsonxx::Object InstanceToJson(TypesCollection& types)
-    {
-        jsonxx::Object ret;
-        InstanceToJsonBase(mTlv, ret);
-
-        jsonxx::Object properties;
-        PropertiesToJson(types, properties);
-        ret << "properties" << properties;
-        return ret;
-    }
-
-    virtual void PropertiesToJson(TypesCollection& types, jsonxx::Object& properties) = 0;
-
-private:
-    struct PropertyInfo
-    {
-        std::string mName;
-        std::string mTypeName;
-    };
-    std::map<void*, PropertyInfo> mInfo;
-    std::string mName;
-    Path_TLV& mTlv;
-protected:
-    std::string mDescription;
-};
-
-#define ADD_PROP(name, type) AddProperty(name, globalTypes.TypeName(typeid(type)), &type)
-
-#define WRITE_BASIC(field) WriteBasicType(field, properties)
-#define WRITE_ENUMS(field) WriteEnumValue(types, properties, field)
-
-#define READ_BASIC(field) ReadBasicType(field, properties)
-#define READ_ENUMS(field) ReadEnumValue(types, field, properties)
-
-namespace Editor
-{
-    class Path_Hoist : public IMapObject
-    {
-    public:
-        Path_Hoist(TypesCollection& globalTypes) : IMapObject("Hoist", mData)
-        {
-            ADD_PROP("HoistType", mData.field_10_type);
-            ADD_PROP("HoistEdgeType", mData.field_12_edge_type);
-            ADD_PROP("Id", mData.field_14_id);
-            ADD_PROP("Scale", mData.field_16_scale);
-        }
-
-        void PropertiesFromJson(TypesCollection& types, jsonxx::Object& properties) override
-        {
-            READ_ENUMS(mData.field_10_type);
-            READ_ENUMS(mData.field_12_edge_type);
-            READ_BASIC(mData.field_14_id);
-            READ_ENUMS(mData.field_16_scale);
-        }
-
-        void PropertiesToJson(TypesCollection& types, jsonxx::Object& properties) override
-        {
-            WRITE_ENUMS(mData.field_10_type);
-            WRITE_ENUMS(mData.field_12_edge_type);
-            WRITE_BASIC(mData.field_14_id);
-            WRITE_ENUMS(mData.field_16_scale);
-        }
-
-    private:
-        ::Path_Hoist mData = {};
-    };
-}
 
 int main(int argc, char* argv[])
 {
-    TypesCollection globalTypes;
 
-    Editor::Path_Hoist eph(globalTypes);
-    auto obj = eph.InstanceToJson(globalTypes);
-    auto v = obj.json();
-    LOG_INFO(v);
-
-    Editor::Path_Hoist eph2(globalTypes);
-    eph2.InstanceFromJson(globalTypes, obj);
+    auto ret = AliveAPI::ExportPathBinaryToJson("Output.json", "C:\\GOG Games\\Abes Exoddus\\MI.LVL", 1);
 
 
 
-    auto obj2 = eph.StructureToJson();
-    auto v2 = obj2.json();
-    LOG_INFO(v2);
-
-    auto obj3 = globalTypes.BasicTypesToJson();
-    auto v3 = obj3.json();
-    LOG_INFO(v3);
-
-    auto obj4 = globalTypes.EnumsToJson();
-    auto v4 = obj4.json();
-    LOG_INFO(v4);
-
-    // TODO: Upgrading
-
-  
-    // Create "fixed" structure data
-    StructuresAndTypes structuresAndTypes;
-    structuresAndTypes.CreateCurrentVersion();
-
-    JsonDocument doc;
-    doc.mStructureAndTypes.CreateCurrentVersion();
-
-    CameraObject testCam;
-    testCam.mName = "lol cam";
-    testCam.mId = 77;
-
-    MapObject testMapObject;
-    testMapObject.mName = "Hoisty mc hoist face";
-    testMapObject.mObjectStructuresType = "Hoist";
-    testMapObject.mWidth = 20;
-    testMapObject.mHeight = 40;
-    testMapObject.mXPos = 999;
-    testMapObject.mYPos = 1234;
-
-    MapObjectProperty testPropertyInt;
-    testPropertyInt.mName = "Id";
-    testPropertyInt.mIntValue = 7;
-    testPropertyInt.mIsString = false;
-    testMapObject.mProperties.push_back(testPropertyInt);
-
-    MapObjectProperty testPropertyString;
-    testPropertyString.mName = "Scale";
-    testPropertyString.mStringValue = "Full";
-    testPropertyString.mIsString = false;
-    testMapObject.mProperties.push_back(testPropertyString);
-
-
-    testCam.mMapObjects.push_back(testMapObject);
-    doc.mCameras.push_back(testCam);
-
-    doc.Save("lol.json");
-    
-    JsonDocument doc2;
-    doc2.Load("lol.json");
-
-    // Read a json file and use the "fixed" data to decode it
-
-    // Compile the json data to a binary path resource
-
-   // AliveAPI::EnumeratePaths("C:\\GOG Games\\Abes Exoddus\\MI.LVL");
-
-    AliveAPI::ExportPathBinaryToJson("Output.json", "C:\\GOG Games\\Abes Exoddus\\MI.LVL", 1);
-
-
-
-    return 0;
+    return ret.errorCode;
 }
