+++ Source/Tools/alive_api.hpp
@@ -30,9 +30,11 @@ namespace AliveAPI
         int numPaths = 0;
     };
 
-    API_EXPORT int GetApiVersion();
-    API_EXPORT Result ExportPathBinaryToJson(const std::string& jsonOutputFile, const std::string& outputLvlFile, int pathResourceId);
-    API_EXPORT Result UpgradePathJson(const std::string& jsonFile);
-    API_EXPORT Result ImportPathJsonToBinary(const std::string& jsonInputFile, const std::string& inputLvlFile, int pathResourceId, const std::vector<std::string>& lvlResourceSources);
-    API_EXPORT EnumeratePathsResult EnumeratePaths(const std::string& inputLvlFile);
+    API_EXPORT [[nodiscard]] int GetApiVersion();
+    API_EXPORT [[nodiscard]] Result ExportPathBinaryToJson(const std::string& jsonOutputFile, const std::string& inputLvlFile, int pathResourceId);
+    API_EXPORT [[nodiscard]] Result UpgradePathJson(const std::string& jsonFile);
+    API_EXPORT [[nodiscard]] Result ImportPathJsonToBinary(const std::string& jsonInputFile, const std::string& outputLvlFile, int pathResourceId, const std::vector<std::string>& lvlResourceSources);
+    API_EXPORT [[nodiscard]] EnumeratePathsResult EnumeratePaths(const std::string& inputLvlFile);
+
+    // TODO: Camera in/exporting
 }
